test(netcdf): organize the crowded tests/netcdf dir into concern subfolders
Move the 62 flat test modules into eight concern packages, mirroring the
existing lazy/ subpackage, so the directory is navigable:

- unit/       netcdf_unit_*, metadata_unit, models_unit (+_netcdf_unit_helpers)
- cf/         CF-convention tests
- plot/       plotting tests (+_plot_helpers)
- selection/  sel, sel_4d, subset, labeled, reduce
- io/         create/read/write, archive, bytes, pickle, streaming, chunking
- structure/  dimensions, groups, variables, attributes, mdim, band-dim
- spatial/    crop, bbox, y-orientation, geostationary, georef
- lazy/       (existing) + xarray_interop, lazy_get_group

conftest.py stays at tests/netcdf/ so its fixtures inherit into every
subpackage; each new folder gets an __init__.py (dotted-import mode). The
two private helper modules move with their consumers and their import
paths are updated (tests.netcdf.plot._plot_helpers,
tests.netcdf.unit._netcdf_unit_helpers), with the affected import blocks
re-sorted. Core/API/integration modules stay at the top level. Pure
relocation: 1590 tests collect unchanged, no test bodies edited.
