+++ tests/netcdf/cf/test_cf_classify.py
