+++ tests/netcdf/cf/test_cf_coord_attrs.py
