+++ tests/netcdf/io/test_pickle.py
