+++ tests/netcdf/lazy/test_lazy_get_group.py
