+++ tests/netcdf/lazy/test_xarray_interop.py
