+++ tests/netcdf/plot/test_plot_animate.py
@@ -10,8 +10,8 @@ from numpy.testing import assert_array_equal
 
 from pyramids.netcdf import FacetSpec, Selectors
 from pyramids.netcdf.netcdf import NetCDF
-from tests.netcdf._plot_helpers import _make_4d_nc, _make_fake_render
 from tests.netcdf.conftest import make_plot_3d_nc
+from tests.netcdf.plot._plot_helpers import _make_4d_nc, _make_fake_render
 
 pytestmark = pytest.mark.plot
