+++ tests/netcdf/plot/test_plot_facet.py
@@ -9,8 +9,8 @@ import pytest
 
 from pyramids.netcdf import ColourOpts, FacetSpec, Selectors
 from pyramids.netcdf._plot import NetCDFPlot
-from tests.netcdf._plot_helpers import _make_4d_nc, _make_curvilinear_nc
 from tests.netcdf.conftest import make_plot_3d_nc
+from tests.netcdf.plot._plot_helpers import _make_4d_nc, _make_curvilinear_nc
 
 pytestmark = pytest.mark.plot
