+++ tests/netcdf/plot/test_plot_render.py
@@ -12,8 +12,8 @@ from numpy.testing import assert_array_equal
 from pyramids.netcdf import ColourOpts, Selectors
 from pyramids.netcdf._plot import NetCDFPlot
 from pyramids.netcdf.netcdf import NetCDF
-from tests.netcdf._plot_helpers import _make_4d_nc
 from tests.netcdf.conftest import make_plot_3d_nc
+from tests.netcdf.plot._plot_helpers import _make_4d_nc
 
 pytestmark = pytest.mark.plot
