+++ tests/netcdf/selection/test_reduce.py
