+++ tests/netcdf/selection/test_sel.py
