+++ tests/netcdf/selection/test_sel_4d.py
