+++ tests/netcdf/selection/test_subset.py
