+++ tests/netcdf/spatial/test_crop_cutline_srs.py
