+++ tests/netcdf/spatial/test_crop_nonspatial_aux.py
