+++ tests/netcdf/spatial/test_crop_string_aux.py
