+++ tests/netcdf/spatial/test_geostationary_geotransform.py
