+++ tests/netcdf/spatial/test_get_variable_georef_no_gdal_error.py
