+++ tests/netcdf/spatial/test_netcdf_bbox.py
