+++ tests/netcdf/spatial/test_y_orientation.py
