+++ tests/netcdf/structure/test_add_variable_metadata.py
