+++ tests/netcdf/structure/test_band_dim_view.py
