+++ tests/netcdf/structure/test_global_attributes.py
