+++ tests/netcdf/structure/test_groups.py
