+++ tests/netcdf/structure/test_rename_variable.py
