+++ tests/netcdf/unit/test_netcdf_unit_properties.py
@@ -8,8 +8,8 @@ import geopandas as gpd
 import pytest
 from shapely.geometry import box
 
-from tests.netcdf._netcdf_unit_helpers import _make_3d_nc
 from tests.netcdf.conftest import make_2d_nc
+from tests.netcdf.unit._netcdf_unit_helpers import _make_3d_nc
 
 pytestmark = pytest.mark.core
