+++ tests/netcdf/unit/test_netcdf_unit_read.py
@@ -9,8 +9,8 @@ import pytest
 from osgeo import gdal
 
 from pyramids.netcdf.netcdf import Container, NetCDF
-from tests.netcdf._netcdf_unit_helpers import _make_3d_nc
 from tests.netcdf.conftest import make_2d_nc
+from tests.netcdf.unit._netcdf_unit_helpers import _make_3d_nc
 
 pytestmark = pytest.mark.core
